requirements.txt
pandas
numpy
scikit-learn
pytest
mlflow
pyyaml
dvc
dagshub
dvc-s3